Implemented Level 1
Level 1:
Design and implement a multi-user chat service using socket programming. This
system should consist of a chat server that hosts a group chat; and a chat client that can be executed
by multiple users to interact with the group chat.
The basic functionality provided by the chat service should allow users to:
1) join a group chat,
2) post a message with the identity of the user to the chat (i.e. the name of the user and the message
is posted to the chat)
3) receive and view messages that are posted by all users within the group chat.