Level 2
Completed Level 2

Level 2:
You are asked to secure your chat service to ensure that chat messages are confidential to the
members of the group chat. To achieve this all messages must be encrypted when sent across a
network.
You must use Java secure socket programming to implement this required functionality (other
approaches to encrypting the messages will not be given marks given this requirement).